Update Notice file: add ezGingerbread/CM info
While this project is not CyanogenMod many enhancements have been made possible from CyanogenMod code
Thus its important that the NOTICE.cm files are included and this comment that CyanogenMod
code is part of the project along with my own.

Any individual commits can be placed in other open source projects without including this Notice
if it is not already part of the codebase its imported to.